refactor: address pr feedback; fixing bugs
