+++ components/QuestionRenderer.tsx
@@ -1,24 +1,32 @@
-import { RadioButtonProps, RadioButtons } from '../RadioButtons'
+import { RadioButtonProps, RadioButtons } from './RadioButtons'
 import {
   SatisfactionSlider,
   SatisfactionSliderProps
-} from '../SatisfactionSlider'
-import { TextResponse, TextResponseProps } from '../TextResponse'
+} from './SatisfactionSlider'
+import { TextResponse, TextResponseProps } from './TextResponse'
 
 export type Question = {
-  // TODO: superior type
+  // TODO: superior type. Needs to allow any type of question
   i18n?: Record<string, any>
   title?: string
   type: string
 } & (RadioButtonProps | SatisfactionSliderProps | TextResponseProps)
 
-const renderQuestion = (
-  question: Question,
+const QuestionRenderer = ({
+  question,
+  updateCallback
+}: {
+  question: Question
   updateCallback?: (update: unknown) => void
-) => {
+}) => {
   const { title, type } = question
 
-  const failure = <h2>{type} was misconfigured.</h2>
+  const failure = (
+    <>
+      <h1>Error</h1>
+      <p>{type} was misconfigured.</p>
+    </>
+  )
 
   switch (type) {
     case 'info':
@@ -56,13 +64,15 @@ const renderQuestion = (
     default:
       console.warn(`Invalid question type ${type}`)
       return (
-        <h2>
-          The configuration contains an invalid question type{' '}
-          <pre style={{ display: 'inline' }}>{type || 'undefined'}</pre>. See
-          console for details.
-        </h2>
+        <>
+          <h1>Error</h1>
+          <p>
+            The configuration contains an invalid question type{' '}
+            <code>{type || 'undefined'}</code>. See console for details.
+          </p>
+        </>
       )
   }
 }
 
-export { renderQuestion }
+export { QuestionRenderer }
