+++ package.json
@@ -10,6 +10,7 @@
     "build-storybook": "build-storybook"
   },
   "dependencies": {
+    "lodash.clone": "^4.5.0",
     "next": "12.1.4",
     "next-intl": "^2.7.4",
     "react": "18.0.0",
@@ -27,6 +28,7 @@
     "@storybook/manager-webpack5": "^6.4.22",
     "@storybook/react": "^6.4.22",
     "@storybook/testing-library": "^0.0.10",
+    "@types/lodash.clone": "^4.5.7",
     "@types/node": "17.0.23",
     "@types/react": "17.0.43",
     "@types/react-dom": "17.0.14",
